DRND-207 browser compatibility check add support for JS files
(cherry picked from commit 76b9e0c4541af17b5fc87fce52ad58413c8037dd)
