--- browserCompatibilityTests/testCss.js
@@ -1,109 +1,0 @@
-// To determine if various browsers support the page's styling,
-// all css files will be run trough postcss-preset-env
-// and checked if the css needs any modification (vendor prefix, polyfill, converting syntax),
-// these changes will be printed out in the cli.
-// This still won't guarantee that the browser even when supporting some css properties,
-// will render them correct. As there is always browser bugs or ambiguity in the spec.
-// run this node script from command line from root folder
-
-/* eslint-disable no-console */
-/* eslint-disable no-await-in-loop */
-/* eslint-disable no-restricted-syntax */
-const { readFile } = require('fs');
-const { promisify } = require('util');
-const { basename } = require('path');
-
-const globby = require('globby');
-const postcss = require('postcss');
-const postcssPresetEnv = require('postcss-preset-env');
-const Diff = require('diff');
-
-const readFileAsync = promisify(readFile);
-
-// most common browsers by usage
-const BROWSERS = [
-  ['last 2 chrome versions', 'last 2 chromeAndroid versions'],
-  ['last 2 safari versions', 'last 2 ios_saf versions'],
-  'last 2 firefox versions',
-  'last 2 samsung versions',
-];
-
-// text color for cmd log
-const COLORS = {
-  reset: '\x1b[0m',
-  dim: '\x1b[2m',
-  red: '\x1b[31m',
-  green: '\x1b[32m',
-  gray: '\x1b[90m',
-};
-
-(async () => {
-  // all css files which will be used on the client. e.g. css for test is ignored
-  const cssFiles = await globby(['styles/*.css', 'blocks/**/*.css']);
-  for (const cssFile of cssFiles) {
-    let filenamePrinted = false;
-    const css = await readFileAsync(cssFile, 'utf8');
-
-    for (const browser of BROWSERS) {
-      let browserNamePrinted = false;
-
-      const result = await postcss([
-        postcssPresetEnv({
-          browsers: browser,
-          enableClientSidePolyfills: true,
-          // don't remove any unnecessary vendor prefixes,
-          // as they might still be necessary for other browser not in the current browser item
-          autoprefixer: { remove: false },
-        }),
-      ]).process(css, { from: undefined });
-
-      result.warnings().forEach((warn) => {
-        console.warn(warn.toString());
-      });
-
-      // eslint-disable-next-line no-continue
-      if (css === result.css) continue; // css didn't need any modification to support the browsers
-      if (!filenamePrinted) {
-        console.log('\n########################################################');
-        console.log(basename(cssFile));
-        console.log('########################################################');
-        filenamePrinted = true;
-      }
-      if (!browserNamePrinted) {
-        console.log('\n\n======================================================');
-        console.log(browser);
-        console.log('======================================================\n');
-        browserNamePrinted = true;
-      }
-
-      const diff = Diff.diffCss(css, result.css);
-
-      const output = [];
-
-      diff.forEach((part) => {
-        // use the common git diff colors: green for added, red for removed to the console output
-        // eslint-disable-next-line no-nested-ternary
-        const color = part.added
-          ? COLORS.green
-          : part.removed
-            ? COLORS.red
-            : COLORS.gray;
-        let { value } = part;
-        // only show couple of lines of the original css when not modified
-        if (!part.added && !part.removed) {
-          const lines = value.split('\n');
-          if (lines.length > 10) {
-            value = [
-              ...lines.slice(0, 5),
-              '\n...\n',
-              ...lines.slice(-5),
-            ].join('\n');
-          }
-        }
-        output.push(color, value);
-      });
-
-      console.log(output.join(''), COLORS.reset);
-    }
-  }
-})();
